[FEATURE] Matrix: Possiblity to set a fixed width for first Column
It is not possible to set a fixed width for the first column (row name) of matrix questions. The calculation of the width of this column depends on the amount of answer columns, which makes the first column width different for questions with different column count.

This patch adds the possibility to set the width of the first column by TypoScript or in the Flexform

Fixes: #8850

git-svn-id: https://svn.typo3.org/TYPO3v4/Extensions/pbsurvey/trunk@66419 735d13b6-9817-0410-8766-e36946ffe9aa