[BUGFIX] Survey starts at first question without answer when using start point after leaving
When using the option "Begin stage when restarting" > "At the point where user left the survey" and having a "Page Break & Conditions" as first item in the survey, the survey starts one page too far when the user returns. It shows the page with the first unanswered question, not the page with the last answered question.

This patch fixes this issue by taking into account there can be a "Page Break" as the first item in the survey

Fixes: #41519

git-svn-id: https://svn.typo3.org/TYPO3v4/Extensions/pbsurvey/trunk@66525 735d13b6-9817-0410-8766-e36946ffe9aa