[FEATURE] Add option to have updateable single response survey only before it has finished
There is the option to have an updateable single response survey. Even when the user finished the survey, it is still possible to update it. There should be an option to stop access to the survey when the user has finished it, but is still updateable when he did not finish it.

This patch fixes this issue

Fixes: #42150

git-svn-id: https://svn.typo3.org/TYPO3v4/Extensions/pbsurvey/trunk@67294 735d13b6-9817-0410-8766-e36946ffe9aa